Conflicts:
	activerecord/CHANGELOG
	activerecord/lib/active_record/associations/class_methods/join_dependency.rb
	activerecord/lib/active_record/associations/class_methods/join_dependency/join_association.rb
	activerecord/lib/active_record/associations/has_many_through_association.rb

+++ railties/lib/rails/generators/migration.rb
@@ -52,7 +52,7 @@ module Rails
 
         destination = self.class.migration_exists?(migration_dir, @migration_file_name)
 
-        if behavior == :invoke
+        if !(destination && options[:skip]) && behavior == :invoke
           if destination && options.force?
             remove_file(destination)
           elsif destination
